[Web] Refactors the old Frontend controller in BackOffice
