+++ src/CriThink.Server/CriThink.Server.Web/Views/BackOffice/Index.cshtml
@@ -1,0 +1,27 @@
+@{
+    Layout = "_Layout";
+    ViewData["Title"] = "Welcome";
+}
+<div class="bg-white p-4 shadow text-center">
+    <img src="logo.svg" width="250" alt="Crithink Logo">
+    <!-- (START) INPUT SECTION -->
+    <form class="p-3">
+        <div class="input-group mb-3">
+            <div class="input-group-prepend">
+                <span class="input-group-text"><i class="fas fa-user"></i></span>
+            </div>
+            <input id="inputemail" type="email" class="form-control" aria-describedby="emailHelp" placeholder="Username or Email">
+        </div>
+        <div class="input-group ">
+            <div class="input-group-prepend">
+                <span class="input-group-text"><i class="fas fa-key"></i></span>
+            </div>
+            <input id="inputpwd" type="password" class="form-control" aria-describedby="emailHelp" placeholder="Password">
+        </div>
+    </form>
+    <a id="btn-login" class="btn btn-purple">Log In</a>
+    <!-- (END) INPUT SECTION -->
+    <div id="errordiv" class="alert alert-danger mt-3 d-none">
+        <span><i class="fas fa-exclamation-circle"></i> Input not valid</span>
+    </div>
+</div>
